fix(core): fix pretty-quick pre-commit hook
Recently `pretty-quick` was adopted for the pre-commit hook. The pattern used in `scripts/format.sh` was not ported to this new utility and is causing issues. In `scripts/format.sh` all HTML files are ignored, but the lack of the appropriate pattern being passed into the pre-commit causes the `packages/workspace/src/core/dep-graph/dep-graph.html` file to fail the `pretty-quick` check.